Update README.md
This PowerBI report is structured as follows:

First, the data is cleaned, selecting the correct data type and dealing with null values and errors, adding the tables and columns needed for further study.

On the first page, we have an overview by brand, which includes the brands with the best average of the most effective boards in powder (virgin snow). In addition, there is an effectiveness/price ratio and the average level of the brand's models.

On the second page, we have a study by brand of the different board models, which includes their effectiveness in powder, as well as their effectiveness/price ratio. It also shows how the styles of the different boards are distributed within the same brand.

Finally, on the last page, we have a customised board search engine. Using height, weight, foot size, style, level and price range, it searches for the brand, model and size that best suit the user's needs, as well as their price, displayed in a conditional colour format depending on whether they are cheap, medium or expensive.